+++ metatensor-torch/CHANGELOG.md
@@ -19,7 +19,21 @@ a changelog](https://keepachangelog.com/en/1.1.0/) format. This project follows
 
 ### metatensor-torch C++
 
+#### Added
+
+- Offer serialization functionality as member functions (i.e. `TensorMap.load`)
+  in addition to the existing free standing functions (i.e. `metatensor_torch::load`) (#453)
+- In-memory serialization with `TensorMap.save_buffer`, `TensorMap.load_buffer`,
+  and the respective free standing functions
+
 ### metatensor-torch Python
+
+#### Added
+
+- Offer serialization functionality as member functions (i.e. `TensorMap.load`)
+  in addition to the existing free standing functions (i.e. `metatensor.torch.load`) (#453)
+- In-memory serialization with `TensorMap.save_buffer`, `TensorMap.load_buffer`,
+  and the respective free standing functions
 
 ## [Version 0.2.0](https://github.com/lab-cosmo/metatensor/releases/tag/metatensor-torch-v0.2.0) - 2024-01-08